moved circular to it's own module
