+++ blender3d_intergration/trajectories_api/definitions.py
@@ -1,9 +1,9 @@
-from abc import abstractmethod, ABC
+from abc import abstractmethod
 
 from matplotlib import pyplot as plt
 
 from blender3d_intergration.trajectories_api.models import Trajectory, \
-    Coordinates, CircularTrajectoryMotionParameters, CircularTrajectoryMotionInitialParameters, FramesPerSecond
+    Coordinates, FramesPerSecond
 
 
 class TrajectoryCalculator:
@@ -47,21 +47,3 @@ class TrajectoryCalculator:
     @abstractmethod
     def _calculate_coordinates(self, frame: int) -> Coordinates:
         raise NotImplementedError
-
-
-class CircleTrajectory(TrajectoryCalculator, ABC):
-    _initial_motion_parameters: CircularTrajectoryMotionInitialParameters
-    _motion_parameters: CircularTrajectoryMotionParameters
-
-    def __init__(self, initial_motion_parameters: CircularTrajectoryMotionInitialParameters, fps: FramesPerSecond):
-        self._initial_motion_parameters = initial_motion_parameters
-        self._motion_parameters = CircularTrajectoryMotionParameters(self._initial_motion_parameters)
-        self._fps = fps
-        self._trajectory = Trajectory()
-
-    def get_initial_motion_parameters(self) -> CircularTrajectoryMotionInitialParameters:
-        return self._initial_motion_parameters
-
-    def get_motion_parameters(self) -> CircularTrajectoryMotionParameters:
-        return self._motion_parameters
-
